refactor: split crate
