+++ office-crypto-core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "docx-decrypt-rs"
+name = "office-crypto-rs"
 version = "0.1.0"
 edition = "2021"
 
@@ -15,8 +15,5 @@ sha1 = "0.10.6"
 sha2 = "0.10.8"
 aes = "0.8.4"
 ecb = "0.1.2"
-wasm-bindgen = "0.2.92"
-js-sys = "0.3.69"
 
 [lib]
-crate-type = ["cdylib"]
