+++ office-crypto-core/src/lib.rs
@@ -1,20 +1,7 @@
-use wasm_bindgen::{prelude::*, JsValue};
+mod crypto;
+mod errors;
 
-pub mod crypto;
-pub mod errors;
-
-#[wasm_bindgen]
-pub fn decrypt(data: JsValue, password: &str) -> Result<Vec<u8>, JsValue> {
-    let array = js_sys::Uint8Array::new(&data);
-    let bytes: Vec<u8> = array.to_vec();
-
-    let decrypted_bytes = match crate::crypto::decrypt_from_bytes(bytes, password) {
-        Ok(bytes) => bytes,
-        Err(e) => return Err(JsValue::from_str(&e.to_string())),
-    };
-
-    Ok(decrypted_bytes)
-}
+pub use crypto::decrypt_from_bytes;
 
 #[cfg(test)]
 mod test {
