--- office-crypto-core/src/main.rs
@@ -1,5 +1,0 @@
-pub mod crypto;
-pub mod errors;
-fn main() {
-    println!("Hello, world!");
-}
