+++ office-crypto-wasm/Cargo.toml
@@ -6,3 +6,8 @@ edition = "2021"
 
 [lib]
 crate-type = ["cdylib"]
+
+[dependencies]
+js-sys = "0.3.69"
+office-crypto-rs = { path = "../office-crypto-core" }
+wasm-bindgen = "0.2.92"
